TW-772: Convert Mqtt and gw-cloud as NodePorts
We are not able to generate the mqttData.log if we use IngressController as NodePorts.
Looked at AP, it was generating the metrics but somehow they were not repoted in the mqttData.log file in the GW-sync pod.
Changing the MQTT and GW-SYNC to NodePort service and then changing the LoadBalancers to use TargetGroup relative to these NodePorts
fixed the issue. Still not clear on the reason.